Arizona jurisdiction; tests for AZ, IN, IA, MT
