+++ tests/load_all_from_repo.py
@@ -82,5 +82,9 @@ def run2(
 
 
 if __name__ == "__main__":
-    run2(election_jurisdiction_list=[("2020 Primary","Georgia")])
+    run2(election_jurisdiction_list=[
+        ("2020 Primary","Indiana"),
+        ("2018 General","Indiana"),
+        ("2016 General","Indiana"),
+    ])
     exit()
