+++ tests/test_Arizona_2020-Primary.py
@@ -2,17 +2,19 @@ import election_data_analysis as e
 #AZ20 tests
 
 def test_presidential(dbname):
+    # TODO get this contest
     #Contest Not available
     assert True == True
 
-
+# TODO get AZ munger to read write-in votes; then will have to add 451 for Bo 'Heir Archy' Garcia
 def test_statewide_totals(dbname):
     assert(not e.data_exists("2020 Primary","Arizona",dbname=dbname) or e.contest_total(
             "2020 Primary",
             "Arizona",
             "US Senate AZ (Democratic Party)",
-        )
-        == 451 + 665620
+        dbname=dbname,
+    )
+        == 665620
     )
 
 def test_state_senate_totals(dbname):
@@ -20,7 +22,8 @@ def test_state_senate_totals(dbname):
             "2020 Primary",
             "Arizona",
             "AZ Senate District 10 (Republican Party)",
-        )
+        dbname=dbname,
+    )
         == 19891
     )
 
@@ -28,13 +31,20 @@ def test_state_rep_totals(dbname):
     assert (not e.data_exists("2020 Primary","Arizona",dbname=dbname) or e.contest_total(
             "2020 Primary",
             "Arizona",
+            "AZ House District 6 (Democratic Party)",
+            dbname=dbname,
+        )
+        == 24035
+    )
+
+
+def test_congressional_totals(dbname):
+    assert (not e.data_exists("2020 Primary","Arizona",dbname=dbname) or e.contest_total(
+            "2020 Primary",
+            "Arizona",
             "US House AZ District 6 (Democratic Party)",
+            dbname=dbname,
         )
         == 3651 + 29218 + 4592 + 42538
     )
 
-def test_contest_by_vote_type(dbname):
-    assert True == True
-
-def test_totals_match_vote_type(dbname):
-    assert True == True
