+++ tests/test_Indiana_2018-General.py
@@ -1,0 +1,37 @@
+import election_data_analysis as e
+
+
+def test_in_statewide_totals_18(dbname):
+    assert (not e.data_exists("2018 General","Indiana",dbname=dbname) or e.contest_total(
+            "2018 General",
+            "Indiana",
+            "US Senate IN",
+            dbname=dbname,
+        )
+            == 2282565
+    )
+
+
+def test_in_senate_totals_18(dbname):
+    assert (not e.data_exists("2018 General","Indiana",dbname=dbname) or e.contest_total(
+            "2018 General",
+            "Indiana",
+            "IN Senate District 14",
+            dbname=dbname,
+        )
+            == 34542
+    )
+
+
+def test_in_house_totals_18(dbname):
+    assert (not e.data_exists("2018 General","Indiana",dbname=dbname) or e.contest_total(
+            "2018 General",
+            "Indiana",
+            "IN House District 27",
+            dbname=dbname,
+        )
+            == 12238
+    )
+
+# Vote Type not available
+
