+++ tests/test_Iowa_2018-General.py
@@ -2,34 +2,34 @@ import election_data_analysis as e
 #IA18 tests
 
 
-def test_presidential():
-    assert True == True
-
-def test_statewide_totals():
+def test_statewide_totals(dbname):
     assert(not e.data_exists("2018 General","Iowa",dbname=dbname) or e.contest_total(
             "2018 General",
             "Iowa",
             "IA Governor",
-        )
+        dbname=dbname,
+    )
         == 1327638
     )
 
-def test_state_senate_totals():
+def test_state_senate_totals(dbname):
     assert (not e.data_exists("2018 General","Iowa",dbname=dbname) or e.contest_total(
             "2018 General",
             "Iowa",
             "IA Senate District 13",
+            dbname=dbname,
         )
         == 30787
     )
 
-def test_state_rep_totals():
+def test_state_rep_totals(dbname):
     assert (not e.data_exists("2018 General","Iowa",dbname=dbname) or e.contest_total(
             "2018 General",
             "Iowa",
-            "US House IA District 14",
-        )
-        == 1382
+            "IA House District 14",
+        dbname=dbname,
+    )
+        == 8551
     )
 
 # no tests by vote type
