+++ tests/test_Iowa_2020-Primary.py
@@ -1,22 +1,15 @@
 import election_data_analysis as e
 #IA20 tests
 
-def test_presidential(dbname):
-    assert(not e.data_exists("2020 Primary","Iowa",dbname=dbname) or e.contest_total(
-            "2020 Primary",
-            "Iowa",
-            "US President (IA) (Democratic Party)",
-        )
-        == 149973
-    )
-
+# Iowa has Presidential caucuses, not primaries
 
 def test_statewide_totals(dbname):
     assert(not e.data_exists("2020 Primary","Iowa",dbname=dbname) or e.contest_total(
             "2020 Primary",
             "Iowa",
             "US Senate IA (Republican Party)",
-        )
+        dbname=dbname,
+    )
         == 229721
     )
 
@@ -25,7 +18,8 @@ def test_state_senate_totals(dbname):
             "2020 Primary",
             "Iowa",
             "IA Senate District 2 (Republican Party)",
-        )
+        dbname=dbname,
+    )
         == 11583
     )
 
@@ -33,13 +27,10 @@ def test_state_rep_totals(dbname):
     assert (not e.data_exists("2020 Primary","Iowa",dbname=dbname) or e.contest_total(
             "2020 Primary",
             "Iowa",
-            "US House IA District 8 (Democratic Party)",
-        )
+            "IA House District 8 (Democratic Party)",
+            dbname=dbname,
+    )
         == 1382
     )
 
-def test_contest_by_vote_type(dbname):
-    assert True == True
-
-def test_totals_match_vote_type(dbname):
-    assert True == True
+# no vote totals
