+++ tests/test_Montana_2020-Primary.py
@@ -6,6 +6,7 @@ def test_presidential(dbname):
             "2020 Primary",
             "Montana",
             "US President (MT) (Democratic Party)",
+            dbname=dbname,
         )
         == 149973
     )
@@ -16,6 +17,7 @@ def test_statewide_totals(dbname):
             "2020 Primary",
             "Montana",
             "US Senate MT (Republican Party)",
+            dbname=dbname,
         )
         == 219205
     )
@@ -25,6 +27,7 @@ def test_state_senate_totals(dbname):
             "2020 Primary",
             "Montana",
             "MT Senate District 17 (Republican Party)",
+            dbname=dbname,
         )
         == 4896
     )
@@ -33,13 +36,9 @@ def test_state_rep_totals(dbname):
     assert (not e.data_exists("2020 Primary","Montana",dbname=dbname) or e.contest_total(
             "2020 Primary",
             "Montana",
-            "US House MT District 4",
-        )
-        == 1977
+            "MT House District 4 (Republican Party)",
+            dbname=dbname,
+    )
+        == 2782
     )
 
-def test_contest_by_vote_type(dbname):
-    assert True == True
-
-def test_totals_match_vote_type(dbname):
-    assert True == True
