+++ tests/test_dataloader.py
@@ -1,83 +1,6 @@
 import election_data_analysis as e
 import pytest
 
-# #### Tests start below #### #
-# For each state, run at least 6 tests:
-# 1. Presidential
-# 2. One statewide chosen at random
-# 3. One senate
-# 4. One rep
-# 5. If vote type is available, slice one of the above by vote type
-# 6. If vote type is available, check that totals match vote type sums
-
-
-#constants
-
-
-### NC dataloading tests ###
-#NC16 tests
-
-def test_nc_presidential_16(dbname):
-    assert(not e.data_exists("2016 General", "North Carolina", dbname=dbname) or e.contest_total(
-            "2016 General",
-            "North Carolina",
-            "US President (NC)",
-            dbname=dbname,
-        )
-            == 4741564
-    )
-
-
-
-def test_nc_statewide_totals_16(dbname):
-    assert (not e.data_exists("2016 General","North Carolina",dbname=dbname) or e.contest_total(
-            "2016 General",
-            "North Carolina",
-            "NC Treasurer",
-            dbname=dbname,
-        )
-            == 4502784
-    )
-
-
-def test_nc_senate_totals_16(dbname):
-    assert (not e.data_exists("2016 General","North Carolina",dbname=dbname) or e.contest_total(
-            "2016 General",
-            "North Carolina",
-            "US Senate NC",
-            dbname=dbname,
-        )
-            == 4691133
-    )
-
-
-def test_nc_rep_16(dbname):
-    assert (not e.data_exists("2016 General","North Carolina",dbname=dbname) or e.contest_total(
-            "2016 General",
-            "North Carolina",
-            "US House NC District 4",
-            dbname=dbname,
-        )
-            == 409541
-    )
-
-
-def test_nc_contest_by_vote_type_16(dbname):
-    assert ( not e.data_exists("2016 General","North Carolina", dbname=dbname) or
-            e.count_type_total(
-            "2016 General",
-            "North Carolina",
-"US House NC District 4",
-            "absentee-mail",
-            dbname=dbname,
-        )
-            == 20881
-    )
-
-
-def test_nc_totals_match_vote_type_16(dbname):
-    assert (not e.data_exists("2016 General","North Carolina", dbname=dbname) or 
-            e.check_totals_match_vote_types("2016 General","North Carolina" ,dbname=dbname) == True)
 
 
 #NC18 tests
@@ -349,123 +272,6 @@ def test_pa_house_totals_20(dbname):
             == 6327
     )
 
-
-### Georgia Data Loading Tests ###
-#GA16
-
-
-#GA18
-
-#GA16
-
-### South Carolina Data Loading Tests ###
-#SC20 test
-
-### Indiana Data Loading Tests ###
-#IN16 test
-
-#IN18 test
-
-def test_in_presidential_18(dbname):
-    assert True == True
-
-
-def test_in_statewide_totals_18(dbname):
-    assert (not e.data_exists("2018 General","Indiana",dbname=dbname) or e.contest_total(
-            "2018 General",
-            "Indiana",
-            "US Senate IN",
-            dbname=dbname,
-        )
-            == 2282565
-    )
-
-
-def test_in_senate_totals_18(dbname):
-    assert (not e.data_exists("2018 General","Indiana",dbname=dbname) or e.contest_total(
-            "2018 General",
-            "Indiana",
-            "IN Senate District 14",
-            dbname=dbname,
-        )
-            == 34542
-    )
-
-
-def test_in_house_totals_18(dbname):
-    assert (not e.data_exists("2018 General","Indiana",dbname=dbname) or e.contest_total(
-            "2018 General",
-            "Indiana",
-            "IN House District 27",
-            dbname=dbname,
-        )
-            == 12238
-    )
-
-
-def test_in_contest_by_vote_type_18(dbname):
-    # Vote type not available
-    assert True == True
-
-
-def test_in_totals_match_vote_type_18(dbname):
-    # Vote type not available
-    assert True == True
-
-#IN20 test
-
-def test_in_presidential_20(dbname):
-    assert (not e.data_exists("2020 Primary","Indiana",dbname=dbname) or e.contest_total(
-            "2020 Primary",
-            "Indiana",
-            "US President (IN)",
-            dbname=dbname,
-        )
-            == 1047173
-    )
-
-
-def test_in_statewide_totals_20(dbname):
-    assert (not e.data_exists("2020 Primary","Indiana",dbname=dbname) or e.contest_total(
-            "2020 Primary",
-            "Indiana",
-            "IN Governor",
-            dbname=dbname,
-        )
-            == 932726
-    )
-
-
-def test_in_senate_totals_20(dbname):
-    assert (not e.data_exists("2020 Primary","Indiana",dbname=dbname) or e.contest_total(
-            "2020 Primary",
-            "Indiana",
-            "IN Senate District 50",
-            dbname=dbname,
-        )
-            == 6860
-    )
-
-
-def test_in_house_totals_20(dbname):
-    assert (not e.data_exists("2020 Primary","Indiana",dbname=dbname) or e.contest_total(
-            "2020 Primary",
-            "Indiana",
-            "IN House District 3",
-            dbname=dbname,
-        )
-            == 7975
-    )
-
-
-def test_in_contest_by_vote_type_20(dbname):
-    # Vote type not available
-    assert True == True
-
-
-def test_in_totals_match_vote_type_20(dbname):
-    # Vote type not available
-    assert True == True
 
 
 ### Arkansas Data Loading Tests ###
